Fix things for linux
