+++ internal/trust/trust_darwin.go
@@ -3,7 +3,10 @@
 package trust
 
 import (
+	"encoding/base64"
 	"fmt"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/kyma-project/cli/internal/cli"
 	"github.com/kyma-project/cli/internal/kube"
@@ -28,6 +31,28 @@ func (k keychain) Certificate() ([]byte, error) {
 		return certificateFromSecret(k)
 	}
 	return certificateFromConfigMap(k)
+}
+
+func certificateFromConfigMap(k keychain) ([]byte, error) {
+	cm, err := k.k8s.Static().CoreV1().ConfigMaps(k.source.namespace).Get(k.source.name, metav1.GetOptions{})
+	if err != nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("\nCould not retrieve the Kyma root certificate. Follow the instructions to import it manually:\n-----\n%s-----\n", k.Instructions()))
+	}
+
+	decodedCert, err := base64.StdEncoding.DecodeString(cm.Data["global.ingress.tlsCrt"])
+	if err != nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("\nCould not retrieve the Kyma root certificate. Follow the instructions to import it manually:\n-----\n%s-----\n", k.Instructions()))
+	}
+
+	return decodedCert, nil
+}
+
+func certificateFromSecret(k keychain) ([]byte, error) {
+	secret, err := k.k8s.Static().CoreV1().Secrets(k.source.namespace).Get(k.source.name, metav1.GetOptions{})
+	if err != nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("\nCould not retrieve the Kyma root certificate. Follow the instructions to import it manually:\n-----\n%s-----\n", k.Instructions()))
+	}
+	return secret.Data["tls.crt"], nil
 }
 
 func (k keychain) StoreCertificate(file string, i Informer) error {
